+++ src/native/libs/System.Security.Cryptography.Native.Android/pal_trust_manager.h
@@ -7,4 +7,4 @@ PALEXPORT void AndroidCryptoNative_RegisterRemoteCertificateValidationCallback(R
 jobjectArray InitTrustManagersWithDotnetProxy(JNIEnv* env, intptr_t sslStreamProxyHandle);
 
 JNIEXPORT jboolean JNICALL Java_net_dot_android_crypto_DotnetProxyTrustManager_verifyRemoteCertificate(
-    JNIEnv *env, jobject thisHandle, intptr_t sslStreamProxyHandle);
+    JNIEnv *env, jobject thisHandle, jlong sslStreamProxyHandle);
